refactor: refactor conversation settings component
